+++ sdk/storage/Azure.Storage.DataMovement.Blobs/src/DataMovementBlobsExtensions.cs
@@ -468,6 +468,8 @@ namespace Azure.Storage.DataMovement.Blobs
             bool overwrite,
             StorageResourceItemProperties sourceProperties)
         {
+            AccessTier? accessTier = GetAccessTier(options, sourceProperties?.RawProperties);
+            PremiumPageBlobAccessTier? pageBlobAccessTier = accessTier.ToPremiumPageBlobAccessTier();
             return new PageBlobCreateOptions()
             {
                 SequenceNumber = options?.SequenceNumber,
@@ -482,7 +484,7 @@ namespace Azure.Storage.DataMovement.Blobs
                     TagConditions = options?.DestinationConditions?.TagConditions,
                     LeaseId = options?.DestinationConditions?.LeaseId,
                 },
-                PremiumPageBlobAccessTier = options?.AccessTier,
+                PremiumPageBlobAccessTier = GetAccessTier(options, sourceProperties?.RawProperties).ToPremiumPageBlobAccessTier(),
             };
         }
 
@@ -683,5 +685,20 @@ namespace Azure.Storage.DataMovement.Blobs
                 : properties?.TryGetValue(DataMovementConstants.ResourceProperties.Metadata, out object metadataObject) == true
                     ? (Metadata)metadataObject
                     : default;
+
+        // Convert AccessTier to PremiumPageBlobAccessTier
+        // As long as it works. Do not set if the AccessTier is a BlockBlob tier
+        private static PremiumPageBlobAccessTier? ToPremiumPageBlobAccessTier(this AccessTier? accessTier)
+        {
+            if (accessTier != default &&
+                accessTier != AccessTier.Hot &&
+                accessTier != AccessTier.Cool &&
+                accessTier != AccessTier.Archive &&
+                accessTier != AccessTier.Cold)
+            {
+                return new PremiumPageBlobAccessTier(accessTier.ToString());
+            }
+            return default;
+        }
     }
 }